requirements.txt
numpy
pandas
scikit-learn
plotly
matplotlib
streamlit
prophet
xgboost
pyyaml